+++ main.py
@@ -16,111 +16,143 @@ change_to = direction
 # Setup fruit
 fruit.init()
 
-# Returns `default_value` if the key doesn't exist
-game_mode = os.getenv('GAME_MODE', 'play')
+# def get_keys():
 
-if game_mode == "play":
-	auto_pilot = False
-elif game_mode == "auto":
-	auto_pilot == True
-else:
-	print("Invalid arg passed ad GAME_MODE : must be auto or play")
-	quit() # quit the program
+# 	if game_mode == "auto":
+# 		auto_pilot = True
+# 	elif game_mode == "std":
+# 		auto_pilot = False
+
+# 	# handling key events
+# 	for event in pygame.event.get():
+# 		if event.type == pygame.KEYDOWN:
+# 			if event.key == pygame.K_ESCAPE:
+# 				game_mode = None
+# 				return
+# 			if auto_pilot is False:
+# 				if event.key == pygame.K_UP or pygame.K_w:
+# 					change_to = 'UP'
+# 				if event.key == pygame.K_DOWN or pygame.K_s:
+# 					change_to = 'DOWN'
+# 				if event.key == pygame.K_LEFT or pygame.K_a:
+# 					change_to = 'LEFT'
+# 				if event.key == pygame.K_RIGHT or pygame.K_d:
+# 					change_to = 'RIGHT'
+# 			else:
+# 				# asking to the autopilot routine!
+# 				change_to = autopilot.find_path(direction)
+# 	return
 
 async def main():
 
 	# Initialising game
-	game_window = game.init()
-
-	# Main Function
+	screen = game.init()
 	while True:
 
-		global change_to, direction, fruit_position, fruit_spawn, score
-	
-		if auto_pilot is False:
-			# handling key events
-			for event in pygame.event.get():
-				if event.type == pygame.KEYDOWN:
-					if event.key == pygame.K_UP:
-						change_to = 'UP'
-					if event.key == pygame.K_DOWN:
-						change_to = 'DOWN'
-					if event.key == pygame.K_LEFT:
-						change_to = 'LEFT'
-					if event.key == pygame.K_RIGHT:
-						change_to = 'RIGHT'
+		global change_to, direction, fruit_position, fruit_spawn, score, game_mode
+
+		# for event in pygame.event.get():
+		# 	if event.type == pygame.QUIT:
+		# 		pygame.quit()
+		# 		quit()
+
+		if game_mode is None:
+			game_mode = game.menu(screen)
 		else:
-			# asking to the autopilot routine!
-			change_to = autopilot.find_path(direction)	
+			if game_mode == "auto":
+				auto_pilot = True
+			elif game_mode == "std":
+				auto_pilot = False
+			print(game_mode)
+			print(auto_pilot)
 
-		# We don't want the new direction to be the
-		# opposite of the current one
-		if change_to == 'UP' and direction != 'DOWN':
-			direction = 'UP'
-		if change_to == 'DOWN' and direction != 'UP':
-			direction = 'DOWN'
-		if change_to == 'LEFT' and direction != 'RIGHT':
-			direction = 'LEFT'
-		if change_to == 'RIGHT' and direction != 'LEFT':
-			direction = 'RIGHT'
+			# handling key events
+			# get_keys()
+			if auto_pilot is False:
+				for event in pygame.event.get():
+					if event.type == pygame.KEYDOWN:
+						if event.key == pygame.K_UP:# or pygame.K_w:
+							change_to = 'UP'
+						if event.key == pygame.K_DOWN:# or pygame.K_s:
+							change_to = 'DOWN'
+						if event.key == pygame.K_LEFT:# or pygame.K_a:
+							change_to = 'LEFT'
+						if event.key == pygame.K_RIGHT:# or pygame.K_d:
+							change_to = 'RIGHT'
+			else:
+				# asking to the autopilot routine!
+				change_to = autopilot.find_path(direction)
 
-		# Moving the snake
-		if direction == 'UP':
-			snake.position[1] -= 10
-		if direction == 'DOWN':
-			snake.position[1] += 10
-		if direction == 'LEFT':
-			snake.position[0] -= 10
-		if direction == 'RIGHT':
-			snake.position[0] += 10
+			# We don't want the new direction to be the
+			# opposite of the current one
+			if change_to == 'UP' and direction != 'DOWN':
+				direction = 'UP'
+			if change_to == 'DOWN' and direction != 'UP':
+				direction = 'DOWN'
+			if change_to == 'LEFT' and direction != 'RIGHT':
+				direction = 'LEFT'
+			if change_to == 'RIGHT' and direction != 'LEFT':
+				direction = 'RIGHT'
 
-		# Check if the fruit was eaten #TODO
-		snake.move()
+			# Moving the snake
+			if direction == 'UP':
+				snake.position[1] -= 10
+			if direction == 'DOWN':
+				snake.position[1] += 10
+			if direction == 'LEFT':
+				snake.position[0] -= 10
+			if direction == 'RIGHT':
+				snake.position[0] += 10
 
-		if fruit.spawn == False:
-			fruit.spawn = True
-			wall.new_wall()
-			fruit.position = fruit.locate()
-			for block in wall.body:
-				while fruit.position == block:
-					fruit.position = fruit.locate()
+			# Check if the fruit was eaten #TODO
+			snake.move()
+
+			if fruit.spawn == False:
+				fruit.spawn = True
+				wall.new_wall()
+				fruit.position = fruit.locate()
+				for block in wall.body:
+					while fruit.position == block:
+						fruit.position = fruit.locate()
+				
+			# Fill the game background
+			game.fill(screen)
 			
-		# Fill the game background
-		game.fill(game_window)
-		
-		# Move the snake body
-		snake.draw(game_window)
+			# Move the snake body
+			snake.draw(screen)
 
-		# Spawn the fruit randomly
-		fruit.draw(game_window)
-		
-		# Draw the walls
-		wall.draw(game_window)
+			# Spawn the fruit randomly
+			fruit.draw(screen)
+			
+			# Draw the walls
+			wall.draw(screen)
 
-		# Periodic boundary conditions
-		if snake.position[0] < 0:
-			snake.position[0] = game.window_x-10
-		if snake.position[0] > game.window_x-10:
-			snake.position[0] = 0
-		if snake.position[1] < 0:
-			snake.position[1] = game.window_y-10
-		if snake.position[1] > game.window_y-10:
-			snake.position[1] = 0
+			# Periodic boundary conditions
+			if snake.position[0] < 0:
+				snake.position[0] = game.window_x-10
+			if snake.position[0] > game.window_x-10:
+				snake.position[0] = 0
+			if snake.position[1] < 0:
+				snake.position[1] = game.window_y-10
+			if snake.position[1] > game.window_y-10:
+				snake.position[1] = 0
 
-		# Touching the snake body
-		# Implement game over conditions if the snake touches itself #TODO
-		for block in snake.body[1:]:
-			if snake.position == block:
-				game.game_over(game_window)
-		
-		# Touching the wall, game over condition
-		for block in wall.body:
-			if snake.position == block:
-				game.game_over(game_window)
-		
+			# Touching the snake body
+			# Implement game over conditions if the snake touches itself #TODO
+			for block in snake.body[1:]:
+				if snake.position == block:
+					game.game_over(screen)
+			
+			# Touching the wall, game over condition
+			for block in wall.body:
+				if snake.position == block:
+					game.game_over(screen)
 
 		# Refresh game
-		game.update(game_window)
+		game.update(screen)
 		await asyncio.sleep(0)
 
-asyncio.run(main())
+game_mode = None
+
+if __name__ == "__main__":
+    asyncio.run(main())
